refactor : 타입 분리
