+++ src/components/common/Badge/Badge.tsx
@@ -1,10 +1,6 @@
+import { BadgeProps } from '../../../type';
 import { BadgeContainer } from './Badge.style';
 import { PropsWithChildren } from 'react';
-
-export interface BadgeProps {
-  color: string;
-  bgColor: string;
-}
 
 const Badge: React.FC<PropsWithChildren<BadgeProps>> = ({
   children,
